+++ packages/stream_core_flutter/lib/src/theme/stream_theme_extensions.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/widgets.dart';
 
+import 'components/stream_app_bar_theme.dart';
 import 'components/stream_audio_waveform_theme.dart';
 import 'components/stream_avatar_theme.dart';
 import 'components/stream_badge_count_theme.dart';
@@ -79,6 +80,9 @@ extension StreamThemeExtension on BuildContext {
 
   /// Returns the [StreamBoxShadow] from the current theme.
   StreamBoxShadow get streamBoxShadow => streamTheme.boxShadow;
+
+  /// Returns the [StreamAppBarThemeData] from the nearest ancestor.
+  StreamAppBarThemeData get streamAppBarTheme => StreamAppBarTheme.of(this);
 
   /// Returns the [StreamAudioWaveformThemeData] from the nearest ancestor.
   StreamAudioWaveformThemeData get streamAudioWaveformTheme => StreamAudioWaveformTheme.of(this);